fix(cuda): ALP patches offset>=1024, PatchesCursor overflow, padding UB
- Fix double-counted offset in ALP scalar_op: subtract chunk_start from
  absolute chunk to get correct index into sliced chunk_offsets array.
  Fixes CUDA_ERROR_ILLEGAL_ADDRESS for slices with offset >= 1024.
- Widen PatchesCursor::remaining from uint8_t to uint32_t to handle
  chunks with >255 patches.
- Zero-initialize GPUPatches before serialization to avoid uninitialized
  padding bytes (technically UB under Rust memory model).
- Remove local FL_CHUNK shadow in execute_output_stage.
- Add test_sliced_alp_with_patches_large_offset (offset=1500).

+++ vortex-cuda/kernels/src/patches.cuh
@@ -110,6 +110,6 @@ public:
 private:
     const uint32_t *indices;
     const T *values;
-    uint8_t remaining;
+    uint32_t remaining;
     uint32_t chunk_base;
 };
